requirements.txt
certifi==2018.11.29
Django==2.1.5
djangorestframework==3.9.1
pytz==2018.9
wincertstore==0.2
